Update esp8266.c
The teacher gave me the module of esp8266 connecting to PIC18LF2550. From your original work, we made some changes to make it correspond to our module. Your function esp8266_ip does not work with us, so we made getIP and esp8266_mac functions to retrieve the IP address and MAC address. 

The problem I am facing right now is the SAPmode function. Its purpose is to set the SSID and other parameters of esp8266 when changed to Soft Access Point mode (SoftAP). In my program, when I run with esp8266_mode(2) and the SAPmode("esp_123","1234567890", 5, 3), the module still shows its default setting (ssid = AI-THINKER-...).